requirements.txt
# MashCad — Dependencies
# Install: pip install -r requirements.txt

# === Core (required) ===
PySide6>=6.5.0
build123d>=0.5.0
loguru>=0.7.0
numpy>=1.24.0

# === 3D Visualization ===
pyvista>=0.42.0
pyvistaqt>=0.11.0
vtk>=9.2.0

# === 2D Geometry ===
shapely>=2.0.0

# === Tessellation ===
ocp-tessellate>=3.0.0

# === Export ===
ezdxf>=1.0.0

# === Scientific (Solver, Mesh ops) ===
scipy>=1.10.0

# === Mesh conversion (optional) ===
pymeshlab
meshlib
pyransac3d
trimesh>=4.0.0

# === Dev / Test ===
pytest>=7.0.0
pytest-qt>=4.2.0
pytest-cov>=4.0.0

# === Build ===
pyinstaller>=6.0.0
